+++ tests/integration/url.spec.js
@@ -11,23 +11,16 @@ describe('url', function () {
       expect(result.ogTitle).to.be.eql('Wikipedia');
       expect(result.ogDescription).to.be.eql('Wikipedia is a free online encyclopedia, created and edited by volunteers around the world and hosted by the Wikimedia Foundation.');
       expect(result.ogLocale).to.be.eql('mul');
-      expect(result.ogImage).to.be.eql([
-        {
-          height: '183',
-          type: 'png',
-          url: 'portal/wikipedia.org/assets/img/Wikipedia-logo-v2.png',
-          width: '200',
-        },
-      ]);
       expect(result.requestUrl).to.be.eql('http://www.wikipedia.org/');
+      expect(result.charset).to.be.eql('utf8');
       expect(result.success).to.be.eql(true);
       expect(result).to.have.all.keys(
         'ogDescription',
-        'ogImage',
         'ogLocale',
         'ogTitle',
         'requestUrl',
         'success',
+        'charset',
       );
       expect(response).to.be.an('object').and.to.not.be.empty;
     });
@@ -42,23 +35,16 @@ describe('url', function () {
       expect(result.ogTitle).to.be.eql('Wikipedia');
       expect(result.ogDescription).to.be.eql('Wikipedia is a free online encyclopedia, created and edited by volunteers around the world and hosted by the Wikimedia Foundation.');
       expect(result.ogLocale).to.be.eql('mul');
-      expect(result.ogImage).to.be.eql([
-        {
-          height: '183',
-          type: 'png',
-          url: 'portal/wikipedia.org/assets/img/Wikipedia-logo-v2.png',
-          width: '200',
-        },
-      ]);
       expect(result.requestUrl).to.be.eql('https://www.wikipedia.org/');
+      expect(result.charset).to.be.eql('utf8');
       expect(result.success).to.be.eql(true);
       expect(result).to.have.all.keys(
         'ogDescription',
-        'ogImage',
         'ogLocale',
         'ogTitle',
         'requestUrl',
         'success',
+        'charset',
       );
       expect(response).to.be.an('object').and.to.not.be.empty;
     });
@@ -73,23 +59,16 @@ describe('url', function () {
       expect(result.ogTitle).to.be.eql('Wikipedia');
       expect(result.ogDescription).to.be.eql('Wikipedia is a free online encyclopedia, created and edited by volunteers around the world and hosted by the Wikimedia Foundation.');
       expect(result.ogLocale).to.be.eql('mul');
-      expect(result.ogImage).to.be.eql([
-        {
-          height: '183',
-          type: 'png',
-          url: 'portal/wikipedia.org/assets/img/Wikipedia-logo-v2.png',
-          width: '200',
-        },
-      ]);
       expect(result.requestUrl).to.be.eql('http://www.wikipedia.org/');
+      expect(result.charset).to.be.eql('utf8');
       expect(result.success).to.be.eql(true);
       expect(result).to.have.all.keys(
         'ogDescription',
         'ogLocale',
-        'ogImage',
         'ogTitle',
         'requestUrl',
         'success',
+        'charset',
       );
       expect(response).to.be.an('object').and.to.not.be.empty;
     });
@@ -104,23 +83,16 @@ describe('url', function () {
       expect(result.ogTitle).to.be.eql('Wikipedia');
       expect(result.ogDescription).to.be.eql('Wikipedia is a free online encyclopedia, created and edited by volunteers around the world and hosted by the Wikimedia Foundation.');
       expect(result.ogLocale).to.be.eql('mul');
-      expect(result.ogImage).to.be.eql([
-        {
-          height: '183',
-          type: 'png',
-          url: 'portal/wikipedia.org/assets/img/Wikipedia-logo-v2.png',
-          width: '200',
-        },
-      ]);
       expect(result.requestUrl).to.be.eql('http://wikipedia.org/');
+      expect(result.charset).to.be.eql('utf8');
       expect(result.success).to.be.eql(true);
       expect(result).to.have.all.keys(
         'ogDescription',
-        'ogImage',
         'ogLocale',
         'ogTitle',
         'requestUrl',
         'success',
+        'charset',
       );
       expect(response).to.be.an('object').and.to.not.be.empty;
     });
@@ -135,23 +107,16 @@ describe('url', function () {
       expect(result.ogTitle).to.be.eql('Wikipedia');
       expect(result.ogLocale).to.be.eql('mul');
       expect(result.ogDescription).to.be.eql('Wikipedia is a free online encyclopedia, created and edited by volunteers around the world and hosted by the Wikimedia Foundation.');
-      expect(result.ogImage).to.be.eql([
-        {
-          height: '183',
-          type: 'png',
-          url: 'portal/wikipedia.org/assets/img/Wikipedia-logo-v2.png',
-          width: '200',
-        },
-      ]);
       expect(result.requestUrl).to.be.eql('http://wikipedia.org/');
+      expect(result.charset).to.be.eql('utf8');
       expect(result.success).to.be.eql(true);
       expect(result).to.have.all.keys(
         'ogDescription',
-        'ogImage',
         'ogLocale',
         'ogTitle',
         'requestUrl',
         'success',
+        'charset',
       );
       expect(response).to.be.an('object').and.to.not.be.empty;
     });