+++ tests/unit/utils.spec.js
@@ -7,129 +7,129 @@ const validateUrl = (urls, valid, message) => {
     it(`${urls[index]} ${message}`, function () {
       const validate = utils.validate(urls[index], 2000);
       if (valid) {
-        return expect(validate.returnInputUrl).to.not.be.empty;
+        return expect(validate.url).to.not.be.empty;
       }
-      return expect(validate.returnInputUrl).to.be.eql(null);
+      return expect(validate.url).to.be.eql(null);
     });
   }
 };
 
 describe('utils', function () {
-  context('validate', function () {
+  describe('validate', function () {
     context('validing URLs', function () {
       validateUrl([
         'foobar.com',
-        'www.foobar.com',
         'foobar.com/',
-        'valid.au',
-        'http://www.foobar.com/',
-        'HTTP://WWW.FOOBAR.COM/',
-        'https://www.foobar.com/',
-        'HTTPS://WWW.FOOBAR.COM/',
-        'http://www.foobar.com:23/',
-        'http://www.foobar.com:65535/',
-        'http://www.foobar.com:5/',
-        'https://www.foobar.com/',
-        'ftp://www.foobar.com/',
-        'http://www.foobar.com/~foobar',
-        'http://user:pass@www.foobar.com/',
-        'http://user:@www.foobar.com/',
-        'http://127.0.0.1/',
-        'http://10.0.0.0/',
-        'http://189.123.14.13/',
-        'http://duckduckgo.com/?q=%2F',
-        'http://foobar.com/t$-_.+!*\'(),',
-        'http://foobar.com/?foo=bar#baz=qux',
-        'http://foobar.com?foo=bar',
-        'http://foobar.com#baz=qux',
-        'http://www.xn--froschgrn-x9a.net/',
-        'http://xn--froschgrn-x9a.com/',
-        'http://foo--bar.com',
-        'http://høyfjellet.no',
-        'http://xn--j1aac5a4g.xn--j1amh',
-        'http://xn------eddceddeftq7bvv7c4ke4c.xn--p1ai',
-        'http://кулік.укр',
-        'test.com?ref=http://test2.com',
-        'http://[FEDC:BA98:7654:3210:FEDC:BA98:7654:3210]:80/index.html',
-        'http://[1080:0:0:0:8:800:200C:417A]/index.html',
-        'http://[3ffe:2a00:100:7031::1]',
-        'http://[1080::8:800:200C:417A]/foo',
         'http://[::192.9.5.5]/ipng',
         'http://[::FFFF:129.144.52.38]:80/index.html',
+        'http://[1080::8:800:200C:417A]/foo',
+        'http://[1080:0:0:0:8:800:200C:417A]/index.html',
         'http://[2010:836B:4179::836B:4179]',
+        'http://[3ffe:2a00:100:7031::1]',
+        'http://[FEDC:BA98:7654:3210:FEDC:BA98:7654:3210]:80/index.html',
+        'http://10.0.0.0/',
+        'http://127.0.0.1/',
+        'http://189.123.14.13/',
+        'http://duckduckgo.com/?q=%2F',
         'http://example.com/example.json#/foo/bar',
+        'http://foo--bar.com',
+        'http://foobar.com?foo=bar',
+        'http://foobar.com/?foo=bar#baz=qux',
+        'http://foobar.com/t$-_.+!*\'(),',
+        'http://foobar.com#baz=qux',
+        'http://høyfjellet.no',
+        'http://user:@www.foobar.com/',
+        'http://user:pass@www.foobar.com/',
+        'http://www.foobar.com:23/',
+        'http://www.foobar.com:5/',
+        'http://www.foobar.com:65535/',
+        'http://www.foobar.com/',
+        'HTTP://WWW.FOOBAR.COM/',
+        'http://www.foobar.com/~foobar',
+        'http://www.xn--froschgrn-x9a.net/',
+        'http://xn------eddceddeftq7bvv7c4ke4c.xn--p1ai',
+        'http://xn--froschgrn-x9a.com/',
+        'http://xn--j1aac5a4g.xn--j1amh',
+        'http://кулік.укр',
+        'https://www.foobar.com/',
+        'https://www.foobar.com/',
+        'HTTPS://WWW.FOOBAR.COM/',
+        'test.com?ref=http://test2.com',
+        'valid.au',
+        'www.foobar.com',
       ], true, 'should be valid');
 
       validateUrl([
-        'http://localhost:3000/',
-        '//foobar.com',
-        'xyz://foobar.com',
-        'invalid/',
-        'invalid.x',
-        'invalid.',
+        '!.foo.com',
         '.com',
-        'http://com/',
+        '',
+        '*.foo.com',
+        '////foobar.com',
+        '//foobar.com',
+        'ftp://www.foobar.com/',
+        'http://*.foo.com',
+        'http:////foobar.com',
+        'http://\n@www.foobar.com/',
         'http://300.0.0.1/',
-        'mailto:foo@bar.com',
-        'rtmp://foobar.com',
-        'http://www.xn--.com/',
-        'http://xn--.com/',
+        'http://com/',
+        'http://example.com.',
+        'http://foobar/ lol/',
+        'http://foobar/? lol',
+        'http://foobar/# lol',
+        'http://localhost:3000/',
+        'http://localhost:61500this is an invalid url!!!!',
+        'http://lol @foobar.com/',
+        'http://lol: @foobar.com/',
+        'http://lol:lol @foobar.com/',
+        'http://lol:lol:lol@foobar.com/',
+        'http://www.-foobar.com/',
+        'http://www.foo_bar.com/',
+        'http://www.foobar-.com/',
         'http://www.foobar.com:0/',
         'http://www.foobar.com:70000/',
         'http://www.foobar.com:99999/',
-        'http://www.-foobar.com/',
-        'http://www.foobar-.com/',
-        'http://foobar/# lol',
-        'http://foobar/? lol',
-        'http://foobar/ lol/',
-        'http://lol @foobar.com/',
-        'http://lol:lol @foobar.com/',
-        'http://lol:lol:lol@foobar.com/',
-        'http://lol: @foobar.com/',
-        'http://www.foo_bar.com/',
         'http://www.foobar.com/\t',
-        'http://\n@www.foobar.com/',
-        '',
-        `http://foobar.com/${new Array(2083).join('f')}`,
-        'http://*.foo.com',
-        '*.foo.com',
-        '!.foo.com',
-        'http://example.com.',
-        'http://localhost:61500this is an invalid url!!!!',
-        '////foobar.com',
-        'http:////foobar.com',
+        'http://www.xn--.com/',
+        'http://xn--.com/',
         'https://example.com/foo/<script>alert(\'XSS\')</script>/',
+        'invalid.',
+        'invalid.x',
+        'invalid/',
+        'mailto:foo@bar.com',
+        'rtmp://foobar.com',
+        'xyz://foobar.com',
+        `http://foobar.com/${new Array(2083).join('f')}`,
       ], false, 'should be invalid');
     });
 
     context('validing Timeouts', function () {
       it('time out is 2000', function () {
         const validate = utils.validate('foobar.com', 2000);
-        expect(validate.returnInputTimeout).to.eql(2000);
+        expect(validate.timeout).to.eql(2000);
       });
       it('timeout is under 2000', function () {
         const validate = utils.validate('foobar.com', 1000);
-        expect(validate.returnInputTimeout).to.eql(1000);
+        expect(validate.timeout).to.eql(1000);
       });
       it('timeout is above 2000', function () {
         const validate = utils.validate('foobar.com', 3000);
-        expect(validate.returnInputTimeout).to.eql(3000);
+        expect(validate.timeout).to.eql(3000);
       });
       it('timeout is a string', function () {
         const validate = utils.validate('foobar.com', '123');
-        expect(validate.returnInputTimeout).to.eql(2000);
+        expect(validate.timeout).to.eql(2000);
       });
       it('timeout is a bool', function () {
         const validate = utils.validate('foobar.com', true);
-        expect(validate.returnInputTimeout).to.eql(2000);
+        expect(validate.timeout).to.eql(2000);
       });
       it('timeout is empty string', function () {
         const validate = utils.validate('foobar.com', '');
-        expect(validate.returnInputTimeout).to.eql(2000);
+        expect(validate.timeout).to.eql(2000);
       });
       it('timeout is missing', function () {
         const validate = utils.validate('foobar.com');
-        expect(validate.returnInputTimeout).to.eql(2000);
+        expect(validate.timeout).to.eql(2000);
       });
     });
   });
@@ -161,6 +161,32 @@ describe('utils', function () {
     it('when type is foo', function () {
       const valid = utils.isImageTypeValid('foo');
       expect(valid).to.eql(false);
+    });
+  });
+
+  describe('isThisANonHTMLUrl', function () {
+    it('when url is type .png', function () {
+      const valid = utils.isThisANonHTMLUrl('www.foo.com/bar.png');
+      expect(valid).to.eql(true);
+    });
+    it('when url is type .html', function () {
+      const valid = utils.isThisANonHTMLUrl('www.foo.com/bar.html');
+      expect(valid).to.eql(false);
+    });
+  });
+
+  describe('removeNestedUndefinedValues', function () {
+    it('when there is no undef values', function () {
+      const object = utils.removeNestedUndefinedValues({ one: 1 });
+      expect(object).to.eql({ one: 1 });
+    });
+    it('when there is undef values', function () {
+      const object = utils.removeNestedUndefinedValues({ one: 1, two: undefined });
+      expect(object).to.eql({ one: 1 });
+    });
+    it('when there is a nested undef value', function () {
+      const object = utils.removeNestedUndefinedValues({ one: 1, two: { three: undefined } });
+      expect(object).to.eql({ one: 1, two: {} });
     });
   });
 });