+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log
 
+## 4.5.0
+- Adding support for custom meta tags you want to scrape
+- If ogs thinks the URL isn't a HTML page, it will return a 'Must scrape an HTML page' error.
+- Updating Dependencies
+
+## 4.4.0
+- Adding support for app links meta data
+- Removed the `withCharset` option, you can use `onlyGetOpenGraphInfo` now if you do not want charset
+- Removed the `runChar` option, this will always be turned on
+- `options.encoding === null` is now deprecated
+- Updating image fallback to only send back valid URLs
+- Updating Dependencies
+
+## 4.3.1
+- Small code clean up and adding tests
+- Updating Dependencies
+
+## 4.3.0
+- Adding support for request headers
+
 ## 4.2.1
 
 - Make sure item.fieldName exists before trying to use it